requirements.txt
beautifulsoup4>=4.3.0
requests>=1.2.3
python-vlc>=1.1.2
